refactor: move to suitable class
